refactor: re arrange folder into libs
re #40
