+++ src/libs/ZipExtractor/IZipExtractor.ts
